Setup GOV.UK notify for password reset
This includes the notify config and functioning password reset email.

Settings will have to be set for the api key and template_id. The
template_id will be for a Nofity template which simply accepts subject
and body and doesn't do additional formatting. This assumes formatting
done using the view templates.