Create foster parent show page
This is intended to be the main page for foster parents. But also local
authority should be able to access this page, perhaps with slightly
different view.

We can choose to not use the foster parent resource for this "homepage"
of foster parents'. Maybe we can later implement various home pages for
different types of users.