Use shortlist filter from Shortlist model
The model is now the persisted filter and always loads from the
Shortlist model.

The implementation using edit + update with redirect is
not ideal, but we need to define more fine grained requirements before
it makes sense to change it. For example whether the filter should
always be persisted or wheter "saving" should be a user action.